requirements.txt
pydantic~=2.11.7
python-dotenv~=1.1.1
protobuf~=6.31.1
google-auth-oauthlib~=1.2.2
google-api-python-client~=2.174.0
streamlit~=1.46.1